backend/requirements.txt
asgiref==3.8.1cryptography==4.2.0Django==5.1.2django-cors-headers==4.5.0djangorestframework==3.15.2dnspython==2.7.0PyJWT==2.9.0pymongo==4.10.1sqlparse==0.5.1tzdata==2024.2